# docs/swagger.yaml
basePath: /
definitions:
  handler.draftItemBody:
    properties:
      content:
        additionalProperties: {}
        type: object
      parent_id:
        type: string
      rank:
        type: string
      type:
        type: string
    type: object
  model.Card:
    properties:
      collectionID:
        type: string
      createdAt:
        type: string
      definition:
        type: string
      id:
        type: string
      image:
        type: string
      position:
        type: integer
      term:
        type: string
      updatedAt:
        type: string
    type: object
  model.Collection:
    properties:
      cards:
        items:
          $ref: '#/definitions/model.Card'
        type: array
      createdAt:
        type: string
      description:
        type: string
      draftID:
        description: non-nil for collections that have a pending item_draft overlay
          (not stored in DB)
        type: string
      exercises:
        items:
          $ref: '#/definitions/model.Exercise'
        type: array
      id:
        type: string
      isPublic:
        type: boolean
      items:
        description: unified content (item-model)
        items:
          $ref: '#/definitions/model.Item'
        type: array
      shareToken:
        description: non-nil when a share link has been generated
        type: string
      testQuestions:
        items:
          $ref: '#/definitions/model.TestQuestion'
        type: array
      title:
        type: string
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  model.Exercise:
    properties:
      CollectionID:
        type: string
      CreatedAt:
        type: string
      Distractors:
        description: 'bank only: extra words for the shared pool'
        items:
          type: string
        type: array
      ID:
        type: string
      Kind:
        description: '"bank" | "choice" | "quiz"'
        type: string
      Options:
        items:
          $ref: '#/definitions/model.TestAnswer'
        type: array
      Position:
        type: integer
      Question:
        description: 'quiz only: a multiple-choice question (a "test" is now a quiz
          exercise).'
        type: string
      Sentences:
        items:
          $ref: '#/definitions/model.ExerciseSentence'
        type: array
      Title:
        type: string
      UpdatedAt:
        type: string
    type: object
  model.ExerciseSentence:
    properties:
      answer:
        description: one word per blank, in order
        items:
          type: string
        type: array
      distractors:
        description: 'choice only: wrong options per blank'
        items:
          items:
            type: string
          type: array
        type: array
      id:
        type: string
      position:
        type: integer
      text:
        type: string
    type: object
  model.Item:
    properties:
      collectionID:
        description: nil = free-floating
        type: string
      content:
        additionalProperties: {}
        description: type-specific fields (JSONB)
        type: object
      createdAt:
        type: string
      id:
        type: string
      parentID:
        description: structural parent (e.g. sentence → exercise)
        type: string
      rank:
        description: fractional index (see internal/rank)
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
  model.TestAnswer:
    properties:
      explanation:
        type: string
      id:
        type: string
      is_correct:
        type: boolean
      position:
        type: integer
      text:
        type: string
    type: object
  model.TestQuestion:
    properties:
      collectionID:
        type: string
      createdAt:
        type: string
      id:
        type: string
      image:
        type: string
      options:
        items:
          $ref: '#/definitions/model.TestAnswer'
        type: array
      position:
        type: integer
      question:
        type: string
      updatedAt:
        type: string
    type: object
  service.DraftCardInput:
    properties:
      definition:
        type: string
      id:
        type: string
      image:
        type: string
      term:
        type: string
    type: object
  service.DraftDiff:
    properties:
      Entries:
        items:
          $ref: '#/definitions/service.DraftDiffEntry'
        type: array
    type: object
  service.DraftDiffEntry:
    properties:
      After:
        allOf:
        - $ref: '#/definitions/model.Item'
        description: staged result; nil when deleted
      Before:
        allOf:
        - $ref: '#/definitions/model.Item'
        description: published item; nil when added
      ItemID:
        type: string
      Status:
        description: '"added" | "changed" | "deleted"'
        type: string
      Type:
        type: string
    type: object
  service.DraftTestInput:
    properties:
      id:
        type: string
      image:
        type: string
      options:
        items:
          $ref: '#/definitions/model.TestAnswer'
        type: array
      question:
        type: string
    type: object
  service.HomeData:
    properties:
      following:
        items:
          $ref: '#/definitions/model.Collection'
        type: array
      own:
        items:
          $ref: '#/definitions/model.Collection'
        type: array
    type: object
  service.PublicCollectionMeta:
    properties:
      cards:
        items:
          $ref: '#/definitions/model.Card'
        type: array
      createdAt:
        type: string
      description:
        type: string
      draftID:
        description: non-nil for collections that have a pending item_draft overlay
          (not stored in DB)
        type: string
      exercises:
        items:
          $ref: '#/definitions/model.Exercise'
        type: array
      followerCount:
        type: integer
      id:
        type: string
      isFollowed:
        type: boolean
      isPublic:
        type: boolean
      items:
        description: unified content (item-model)
        items:
          $ref: '#/definitions/model.Item'
        type: array
      shareToken:
        description: non-nil when a share link has been generated
        type: string
      testQuestions:
        items:
          $ref: '#/definitions/model.TestQuestion'
        type: array
      title:
        type: string
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  service.UpdateDraftReq:
    properties:
      cards:
        items:
          $ref: '#/definitions/service.DraftCardInput'
        type: array
      description:
        type: string
      isPublic:
        type: boolean
      testQuestions:
        items:
          $ref: '#/definitions/service.DraftTestInput'
        type: array
      title:
        type: string
    type: object
  service.UserWithCollections:
    properties:
      collections:
        items:
          $ref: '#/definitions/model.Collection'
        type: array
      id:
        type: string
      name:
        type: string
      picture:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Flashcard study platform API
  title: Cram API
  version: "1.0"
paths:
  /collections:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Collection'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List collections
      tags:
      - collections
    post:
      consumes:
      - application/json
      parameters:
      - description: Collection data
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            is_public:
              type: boolean
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Collection'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a collection
      tags:
      - collections
  /collections/{collectionID}:
    delete:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a collection
      tags:
      - collections
    get:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Collection'
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a collection with cards and test questions
      tags:
      - collections
    put:
      consumes:
      - application/json
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      - description: Collection data
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            is_public:
              type: boolean
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Collection'
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a collection
      tags:
      - collections
  /collections/{collectionID}/cards:
    post:
      consumes:
      - application/json
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      - description: Card data
        in: body
        name: body
        required: true
        schema:
          properties:
            answer:
              type: string
            position:
              type: integer
            question:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Card'
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Add a card to a collection
      tags:
      - cards
  /collections/{collectionID}/cards/{cardID}:
    delete:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      - description: Card ID
        in: path
        name: cardID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a card
      tags:
      - cards
    put:
      consumes:
      - application/json
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      - description: Card ID
        in: path
        name: cardID
        required: true
        type: string
      - description: Card data
        in: body
        name: body
        required: true
        schema:
          properties:
            answer:
              type: string
            position:
              type: integer
            question:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Card'
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a card
      tags:
      - cards
  /collections/{collectionID}/cards/import:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      - description: CSV / JSON / YAML file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              imported:
                type: integer
              skipped:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Bulk import cards from CSV (question;answer) or a JSON/YAML list
      tags:
      - cards
  /collections/{collectionID}/draft:
    delete:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Discard draft and restore the active version
      tags:
      - drafts
    post:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Collection'
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get or create a draft for a collection
      tags:
      - drafts
    put:
      consumes:
      - application/json
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      - description: Draft content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UpdateDraftReq'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Save draft content without publishing
      tags:
      - drafts
  /collections/{collectionID}/draft/diff:
    get:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.DraftDiff'
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Diff the staged draft against the published collection
      tags:
      - drafts
  /collections/{collectionID}/draft/items:
    post:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      - description: Item to stage
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.draftItemBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Item'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Stage a new item into the draft
      tags:
      - drafts
  /collections/{collectionID}/draft/items/{itemID}:
    delete:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      - description: Item ID
        in: path
        name: itemID
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Stage deletion of a single item in the draft
      tags:
      - drafts
    put:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      - description: Item ID
        in: path
        name: itemID
        required: true
        type: string
      - description: New item state
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.draftItemBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Item'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Stage an edit to a single item in the draft
      tags:
      - drafts
  /collections/{collectionID}/draft/items/{itemID}/revert:
    post:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      - description: Item ID
        in: path
        name: itemID
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Revert one item's staged change back to the published state
      tags:
      - drafts
  /collections/{collectionID}/draft/publish:
    post:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Publish the draft as the active version
      tags:
      - drafts
  /collections/{collectionID}/exercises/import:
    post:
      consumes:
      - text/plain
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              imported:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Bulk import exercises from a YAML document
      tags:
      - exercises
  /collections/{collectionID}/follow:
    delete:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Unfollow a collection
      tags:
      - follows
    post:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Follow a collection
      tags:
      - follows
  /collections/{collectionID}/tests:
    post:
      consumes:
      - application/json
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      - description: Question data
        in: body
        name: body
        required: true
        schema:
          properties:
            options:
              items:
                $ref: '#/definitions/model.TestAnswer'
              type: array
            position:
              type: integer
            question:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.TestQuestion'
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Add a test question to a collection
      tags:
      - tests
  /collections/{collectionID}/tests/{tqID}:
    delete:
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      - description: Test question ID
        in: path
        name: tqID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a test question
      tags:
      - tests
    put:
      consumes:
      - application/json
      parameters:
      - description: Collection ID
        in: path
        name: collectionID
        required: true
        type: string
      - description: Test question ID
        in: path
        name: tqID
        required: true
        type: string
      - description: Question data
        in: body
        name: body
        required: true
        schema:
          properties:
            options:
              items:
                $ref: '#/definitions/model.TestAnswer'
              type: array
            position:
              type: integer
            question:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TestQuestion'
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a test question
      tags:
      - tests
  /collections/{collectionID}/tests/import:
    post:
      responses: {}
      summary: Bulk import test questions from CSV or a JSON/YAML list
  /home:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.HomeData'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get home page data (own + followed collections)
      tags:
      - collections
  /public/collections:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.PublicCollectionMeta'
            type: array
      summary: List all public collections with follow metadata
      tags:
      - collections
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.UserWithCollections'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List all users with their public collections
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"